ad-silk-sarees/server/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
email-validator
cloudinary
razorpay
python-multipart
